Refactor Universal Discovery Handler to reduce ACL score and improve maintainability
Summary of changes:
- Refactored `custom_components/meraki_ha/discovery/handlers/universal.py` to resolve High Cognitive Load (ACL) functions.
- Extracted entity instantiation logic into specialized helpers: `_discover_capability`, `_instantiate_entities`, `_instantiate_single_entity`, `_handle_provider_class`, `_handle_special_entities`, and `_handle_default_entity`.
- Flattened complex logic in `discover_entities` and `_handle_default_entity`.
- Updated type hints for better accuracy and type safety.
- Fixed several broken relative imports for the `coordinators` module to resolve `ModuleNotFoundError` during test discovery.

I encountered significant blockers while trying to run tests due to widespread broken imports in the repository (likely from a previous reorganization where `coordinators.py` was moved to a directory). While I fixed several occurrences, many remain, preventing the full test suite from running successfully. However, the refactored `universal.py` was verified to be syntactically correct.
