+++ custom_components/meraki_ha/switch/builders/camera.py
@@ -7,7 +7,7 @@ from homeassistant.components.switch import SwitchEntity
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from ..coordinators import MerakiSwitchCoordinator
+from ...coordinators import MerakiSwitchCoordinator
 from ..camera_controls import AnalyticsSwitch
 
 _LOGGER = logging.getLogger(__name__)
