+++ custom_components/meraki_ha/switch/builders/traffic_shaping.py
@@ -8,7 +8,7 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from ...const_conf import CONF_ENABLE_TRAFFIC_SHAPING
-from ..coordinators import MerakiSwitchCoordinator
+from ...coordinators import MerakiSwitchCoordinator
 from ...core.models.network import MerakiTrafficShaping
 from ..traffic_shaping import MerakiTrafficShapingSwitch
