+++ custom_components/meraki_ha/switch/camera_settings.py
@@ -8,7 +8,7 @@ from homeassistant.components.switch import SwitchEntity
 from homeassistant.const import EntityCategory
 from homeassistant.helpers.device_registry import DeviceInfo
 
-from custom_components.meraki_ha.coordinator import MerakiSwitchCoordinator
+from custom_components.meraki_ha.coordinators import MerakiSwitchCoordinator
 
 from ..core.api.client import MerakiAPIClient
 from ..core.models.device import MerakiDevice
