+++ custom_components/meraki_ha/switch/firewall_rule.py
@@ -10,7 +10,7 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import EntityCategory
 from homeassistant.core import callback
 
-from .coordinators import MerakiSwitchCoordinator
+from ..coordinators import MerakiSwitchCoordinator
 from ..core.entities.meraki_firewall_rule_entity import MerakiFirewallRuleEntity
 from ..core.models.network import MerakiFirewallRule
 from ..core.utils.entity_id_utils import get_firewall_rule_entity_id
