+++ requirements_dev_no_aiodns.txt
@@ -1,0 +1,139 @@
+# requirements_dev.txt
+aiodhcpwatcher
+aiodiscover
+aiofiles==24.1.0
+aiohappyeyeballs
+aiohasupervisor
+aiohttp
+aiohttp-asyncmdnsresolver
+aiohttp-cors
+aiohttp-fast-zlib
+aiooui
+aiosignal
+aiozoneinfo
+annotated-types==0.7.0
+annotatedyaml
+anyio
+astral
+async-timeout
+async_interrupt
+atomicwrites-homeassistant
+attrs
+awesomeversion
+bandit==1.7.9
+bcrypt
+bleak
+bleak-retry-connector
+bluetooth-adapters
+bluetooth-auto-recovery
+bluetooth-data-tools
+boolean.py==5.0
+boto3
+botocore
+btsocket
+build==1.3.0
+CacheControl
+certifi
+cffi==2.0.0
+charset-normalizer
+ciso8601
+codecov==2.1.13
+coverage
+cryptography
+cyclonedx-python-lib==7.6.2
+dbus-fast
+defusedxml==0.7.1
+envs==1.4
+execnet==2.1.2
+filelock==3.20.2
+fnv-hash-fast
+fnvhash
+freezegun
+go2rtc-client==0.2.1
+h11
+habluetooth
+home-assistant-bluetooth
+homeassistant==2024.12.5
+meraki==1.54.0
+mypy
+mypy_extensions==1.1.0
+numpy
+orjson
+packageurl-python==0.17.6
+packaging
+paho-mqtt
+pillow
+pip-api==0.0.34
+pip-requirements-parser==32.0.1
+pip_audit==2.7.3
+pipdeptree
+playwright==1.57.0
+pluggy
+propcache
+psutil
+psutil-home-assistant
+py-serializable==1.1.2
+pycognito==2024.5.1
+pycparser
+pydantic
+pydantic_core
+pyee==13.0.0
+Pygments==2.19.2
+PyJWT==2.10.1
+pylint-per-file-ignores
+pyOpenSSL
+pyparsing==3.2.5
+pyproject_hooks==1.2.0
+pyRFC3339==2.1.0
+PyRIC==0.1.6.3
+pytest-aiohttp
+pytest-asyncio
+pytest-cov
+pytest-github-actions-annotate-failures
+pytest-homeassistant-custom-component==0.13.195
+pytest-picked
+pytest-socket
+pytest-sugar
+pytest-timeout
+pytest-xdist
+PyTurboJPEG==1.7.2
+pytz==2025.2
+PyYAML
+regex
+requests
+requests-mock
+respx
+rich
+ruff==0.5.5
+s3transfer
+securetar
+sentence_stream==1.2.1
+setuptools==78.1.1
+six
+snitun
+sortedcontainers==2.4.0
+SQLAlchemy
+stevedore==5.6.0
+syrupy
+termcolor==3.2.0
+text-unidecode==1.3
+toml==0.10.2
+tqdm
+types-aiofiles==25.1.0.20251011
+typing-inspection==0.4.2
+typing_extensions==4.15.0
+tzdata
+uart-devices
+ulid-transform
+urllib3
+usb-devices
+uv
+voluptuous
+voluptuous-openapi
+voluptuous-serialize
+webencodings==0.5.1
+webrtc-models==0.3.0
+wheel==0.45.1
+yarl
+zeroconf==0.148.0
+google-genai
